requirements.txt
beautifulsoup4==4.11.1
bs4==0.0.1
certifi==2022.6.15
charset-normalizer==2.1.0
idna==3.3
markovify==0.9.4
mypy-extensions==0.4.3
psycopg2-binary==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
pyTelegramBotAPI==4.6.0
requests==2.28.1
soupsieve==2.3.2.post1
tabulate==0.8.10
tomli==2.0.1
types-requests==2.28.0
types-tabulate==0.8.11
types-urllib3==1.26.15
typing_extensions==4.3.0
Unidecode==1.3.4
urllib3==1.26.9
